--- app/src/main/java/com/replica/replicaisland/TouchFilter.kt
@@ -1,28 +1,0 @@
-/*
- * Copyright (C) 2010 The Android Open Source Project
- * Copyright (C) 2025 Jim Andreas kotlin conversion
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *      http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package com.replica.replicaisland
-
-import android.content.Context
-import android.view.MotionEvent
-
-abstract class TouchFilter : BaseObject() {
-    abstract fun updateTouch(event: MotionEvent?)
-    open fun supportsMultitouch(context: Context?): Boolean {
-        return false
-    }
-
-    override fun reset() {}
-}
