+++ app/src/main/java/com/replica/replicaisland/input/SingleTouchFilter.kt
@@ -1,6 +1,7 @@
-package com.replica.replicaisland
+package com.replica.replicaisland.input
 
 import android.view.MotionEvent
+import com.replica.replicaisland.input.TouchFilter
 
 open class SingleTouchFilter : TouchFilter() {
     override fun updateTouch(event: MotionEvent?) {
