+++ app/src/main/java/com/replica/replicaisland/input/TouchFilter.kt
@@ -1,0 +1,14 @@
+package com.replica.replicaisland.input
+
+import android.content.Context
+import android.view.MotionEvent
+import com.replica.replicaisland.BaseObject
+
+abstract class TouchFilter : BaseObject() {
+    abstract fun updateTouch(event: MotionEvent?)
+    open fun supportsMultitouch(context: Context?): Boolean {
+        return false
+    }
+
+    override fun reset() {}
+}
